Added http-proxy allowing paths to proxy out to other servers.
This is useful for when your application isn't in a NodeJS stack. For
example if you're building a site with PHP but you want to use grunt-bbb
for all of the javascript development, you can proxy /api out to your
PHP site's /api. When you're ready to deploy you have completed
javascript code that can run on your PHP domain.